perf(crypto): 8-block parallel AES-XTS; fix kernel stack overflow BSOD
Rewrite XtsVolumeCipher to process 8 AES blocks per batch
(encrypt_blocks -> AES-NI encrypt8) so throughput is bounded by the AES
pipeline rather than per-block latency. Wire encrypt_area/decrypt_area
through the read/write pipeline, the sweep engine, and the bench handler.

Do NOT force -C target-feature=+aes in the driver/loader RUSTFLAGS: it
lets LLVM inline the fully-unrolled AES-NI encrypt8/decrypt8 into the deep
IOCTL/metadata call chain, ballooning kernel stack frames past the small
kernel stack and double-faulting on stack overflow (0x7F). The aes crate
detects AES-NI at runtime (cpufeatures) and still dispatches to its
hardware backend, so keeping it behind that call boundary bounds the
frames. Also mark the per-sector XTS entry points #[inline(never)] and
slim XtsVolumeCipher to two Aes256.

Fix DISPATCH_STACK_SIZE type (usize -> u64) for the callout SIZE_T arg.

Validated: cargo test -p vck-common (31/31) and make test-vm-os-encrypt
(22/22, no BSOD; marker integrity + SWEEP_RESUMED).

+++ sample/windrv/src/lib.rs
@@ -55,7 +55,7 @@ const STATUS_PENDING: NTSTATUS = 0x0000_0103;
 
 // The metadata/crypto path uses a large stack frame; run IOCTL dispatch on an
 // expanded kernel stack so the (deep) storage stack below us has headroom.
-const DISPATCH_STACK_SIZE: usize = 0x8000; // 32 KiB
+const DISPATCH_STACK_SIZE: u64 = 0x8000; // 32 KiB
 
 /// Context passed to the expanded-stack dispatch callout.
 struct ExpandCtx {
